[backend] Change init to remove glob plugin in order to fix vitest coverage
