+++ opencti-platform/opencti-graphql/vitest.config.ts
@@ -2,10 +2,9 @@
 import { defineConfig } from 'vitest/config';
 import graphql from '@rollup/plugin-graphql';
 import type { PluginOption } from 'vite';
-import vitestMigrationPlugin from './builder/plugin/vitestMigrationPlugin';
 
 const buildTestConfig = (include: string[]) => defineConfig({
-  plugins: [graphql() as PluginOption, vitestMigrationPlugin() as PluginOption],
+  plugins: [graphql() as PluginOption],
   test: {
     include,
     testTimeout: 300000,
